Add Json parser and Enums for the data
The Json parser is initialized with a file name
- Reads the file and parses the data to a list of objects
- The date is parsed and stored as datetime
- The numbers are converted to int/float where necessary
- The field isinvalid is removed

Json parser method get() is used to query the data
Examples are provided in the file below the class.

The Enums file contains the enumerations for the data
Because the data is stored as integers these enums provide an easier way
to know what station you can request the data for or what parameters
you wish to look at exactly.